requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
pybind11>=2.11.0
python-dotenv>=1.0.0
django-cors-headers>=4.0.0
